# docs/openapi.yaml
definitions: {}
info:
  description: powered by Flasgger
  termsOfService: /tos
  title: A swagger API
  version: 0.0.1
paths:
  /contacts:
    post:
      parameters:
      - in: body
        name: body
        required: true
        schema:
          properties:
            name:
              example: John Doe
              type: string
            phone:
              example: '+123456789'
              type: string
          type: object
      responses:
        '201':
          description: "\u041A\u043E\u043D\u0442\u0430\u043A\u0442 \u0441\u043E\u0437\
            \u0434\u0430\u043D"
      summary: "\u0421\u043E\u0437\u0434\u0430\u043D\u0438\u0435 \u043D\u043E\u0432\
        \u043E\u0433\u043E \u043A\u043E\u043D\u0442\u0430\u043A\u0442\u0430"
  /contacts/{id}:
    delete:
      parameters:
      - in: path
        name: id
        required: true
        type: integer
      responses:
        '200':
          description: "\u041A\u043E\u043D\u0442\u0430\u043A\u0442 \u0443\u0434\u0430\
            \u043B\u0451\u043D"
        '404':
          description: "\u041A\u043E\u043D\u0442\u0430\u043A\u0442 \u043D\u0435 \u043D\
            \u0430\u0439\u0434\u0435\u043D"
      summary: "\u0423\u0434\u0430\u043B\u0435\u043D\u0438\u0435 \u043A\u043E\u043D\
        \u0442\u0430\u043A\u0442\u0430 \u043F\u043E ID"
    get:
      parameters:
      - in: path
        name: id
        required: true
        type: integer
      responses:
        '200':
          description: "\u0418\u043D\u0444\u043E\u0440\u043C\u0430\u0446\u0438\u044F\
            \ \u043E \u043A\u043E\u043D\u0442\u0430\u043A\u0442\u0435"
        '404':
          description: "\u041A\u043E\u043D\u0442\u0430\u043A\u0442 \u043D\u0435 \u043D\
            \u0430\u0439\u0434\u0435\u043D"
      summary: "\u041F\u043E\u043B\u0443\u0447\u0435\u043D\u0438\u0435 \u0438\u043D\
        \u0444\u043E\u0440\u043C\u0430\u0446\u0438\u0438 \u043E \u043A\u043E\u043D\
        \u0442\u0430\u043A\u0442\u0435 \u043F\u043E ID"
swagger: '2.0'